Add new local error() function encapsulating logging and dispatching.
Fire this from two locations in the HTTP GET handler:
* On promise fulfillment when HTTP status is not 2XX
* on promise failure.
We have only seen HTTP GET errors reported in the second way, but I
have determined experimentally that at least some POST errors (500
Internal Server Error) are reported in the former way. So to be on the
safe side, we need to catch both cases for all four HTTP methods.
*sigh*
See STRIPES-51.