createAction copes with 201 Created responses that have no body
In fact, anything that is not a JSON body is assumed to mean that the
client record is good enough for further consumption, otherwise the
server would have sent us one that it prefers. In this case,
crudActions.createSuccess is passed the client record, which ends up
in successfulMutations.

Fixes STRPCONN-2.