Defensive programming to avoid errors in test environment
`console.log` on an augmented crudActions object invokes `toJSON` on
it. For some reason, in the test environment but not in Stripes, that
function is undefined. Guard against this, and just return the type of
the crudActions object in this case.

Also, in the test environment, attempting to log the value of
`crudActions` itself causes a fatal error -- so we avoid doing that
when we use `console.log` to report an error.

Finally, if an undefined function name (other than `toJSON`) somehow
finds its way into the `augmentWithMetadata` function, then we return
undefined as its value, so that the calling code has some way to
continue rather than crashing out when we try to invoke a
non-function. Why is this necessary? Because for reasons I don't
understand, the test environment causes the crudActions object to
somehow be indexed not by a string but by a Symbol.

All of this should be unnecessary; but the world is not as we would
wish.

Fixes STRPCONN-11.