Use crudName, not stateKey() to generate action-names
I did this purely to simplify the action names, and to clear up a
conceptual confusion: stateKey() is about the key by which a
resource's data is indexed in the Redux store, and properly has no
role in the name of an action.

(More generally, we want to get away from our present tendency to have
information other than action-type encoded in `action.type`. Even the
crudName doesn't really belong there: we should simply include
module-name and resource-name in the event metadata. But that's for
another time.)

Anyway -- to my pleasant surprise, I found that this change (which I
thought would be a no-op) fixed things so that <Users> works correctly
when it has a dataKey. Which I suspect means that it fixes STCON-9.