Support substitution for clientGeneratePk
It's now possible for the `clientGeneratePk` manifest option to take
any of the forms that `path` and other options can take. So for
example, your manifest can specify `clientGeneratePk: '!{foo}'` to use
the value of the `foo` prop. (This is in fact exactly what
`<ControlledVocab>` does to allow its caller to pass in a prop
indicating whether ot not to have the client generate an ID.

As well as adding the line that runs substitute on `clientGeneratePk`,
I had to tweak `substitute` so it passes through boolean values (as
the default value is `true`). While I was there, I made the diagnostic
for an unsupported type a bit more verbose, as that would have helped
me to track down what was otherwise a slightly obscure error thrown by
the test suite.