correctly dispatch PAGE_SUCCESS when paging ends
When paging ends unexpectedly because Okapi goes from returning
`totalRecords=999999999` to e.g. `totalRecords=100`, an extra request
will be triggered that will receive `totalRecords=0`. When this happens,
we need still need to trigger a PAGE_SUCCESS event to indicate that
paging has ended, but we need to do so _without_ resetting the final
record count to 0.

Yeah, it's a bit messy. Additional details in the comments.

Refs STSMACOM-259