provide OkapiLookupResource
A "lookup" resource is a straight clone of `OkapiResource` with one
exception: it provides a limit clause to `GET` queries by default so
that we never again, not EVER, have to worry about accidentally
forgetting to provide one and retrieving only 10 records. It is
hard-coded to retrieve 2000 items.

If your lookup table has more than 2000 items in it, you probably
shouldn't be treating it as an ordinary lookup table because you need to
worry about paging and such like. This isn't for you.