Allow 'headers' to be a function and interpreted
It is conceivable that a request may contain headers that need value
substitution to take place (as per query parameters, path components
etc.). This commit adds that.

The ability for the 'headers' value to be a function as well as an
object, as per 'path' etc. This allows us to access props that have
been passed to the component.

The use case this was implemented for was to allow an external
API key, that was being passed down to a component as a prop, to be used
in an authentication header for an API call being made by the child
component.